Refactored for CD compatibility renamed index.ts to app.ts Configured build script
