+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.ts",
   "scripts": {
-    "start": "ts-node ./src/index.ts",
+    "start": "ts-node ./src/app.ts",
+	"build": "tsc",
     "test": "jest"
   },
   "keywords": [],
