+++ tsconfig.json
@@ -26,7 +26,7 @@
 		}
 	},
 	"files": [
-		"src/index.ts"
+		"src/app.ts"
 	],
 	"include": [
 		"./**/*.ts"
